Add UserStoreScore model
